+++ desktop/src/main/kotlin/android/app/Application.kt
@@ -1,8 +1,9 @@
 package android.app
 
 import android.content.Context
+import android.content.ContextWrapper
 
-class Application : Context("mirah.app") {
+open class Application : ContextWrapper(Context("mirah.app")) {
     companion object {
         var instance: Application = Application()
     }