+++ desktop/src/main/kotlin/android/content/ContextWrapper.kt
@@ -1,0 +1,6 @@
+package android.content
+
+open class ContextWrapper(base: Context?) : Context(base?.packageName ?: "mirah.app") {
+    private var baseContext: Context? = base
+    fun getBaseContext(): Context? = baseContext
+}