+++ desktop/src/main/kotlin/android/content/Intent.kt
@@ -1,0 +1,7 @@
+package android.content
+
+open class Intent {
+    constructor()
+    constructor(action: String)
+    constructor(context: Context?, cls: Class<*>)
+}