+++ desktop/src/main/kotlin/eu/kanade/tachiyomi/AppInfo.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi
+
+object AppInfo {
+    fun getVersionCode(): Int = 1
+    fun getVersionName(): String = "Mirah"
+    fun getSupportedImageMimeTypes(): List<String> = emptyList()
+}