+++ desktop/src/main/kotlin/eu/kanade/tachiyomi/source/model/Page.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.source.model
+
+interface Page {
+    val index: Int
+    var url: String
+    var imageUrl: String?
+}